Move commands into a single module
