--- django_codemod/commands/django_32.py
@@ -1,3 +1,0 @@
-"""Nothing yet!"""
-# This is expected to cover most of the things listed in this section:
-# https://docs.djangoproject.com/en/dev/internals/deprecation/#deprecation-removed-in-3-2
