+++ docs/codemods.rst
@@ -6,51 +6,29 @@ If everything is setup properly, the list of Django Codemods should appear when 
 bash::
 
     > python3 -m libcst.tool list
-    django_30.Django30Command - Resolve deprecations for removals in Django 3.0.
-    django_40.Django40Command - Resolve deprecations for removals in Django 4.0.
+    Django30Command - Resolve deprecations for removals in Django 3.0.
+    Django40Command - Resolve deprecations for removals in Django 4.0.
 
 Codemodders are organised following the Django `deprecation timeline page`_, listing all its deprecations by version.
-
-For instance, to fix a deprecation listed in the 4.0 section, the corresponding codemodder would be found under ``django_40``.
 
 .. _deprecation timeline page: https://docs.djangoproject.com/en/3.0/internals/deprecation/
 
 Django 3.0
 ----------
 
-These codemodders should fix things `removed in Django 3.0`_.
+This command should fix things `removed in Django 3.0`_.
 
 .. _removed in Django 3.0: https://docs.djangoproject.com/en/dev/internals/deprecation/#deprecation-removed-in-3-0
 
-.. automodule:: django_codemod.commands.django_30
-    :members:
-
-Django 3.1
-----------
-
-These codemodders should fix things `removed in Django 3.1`_.
-
-.. _removed in Django 3.1: https://docs.djangoproject.com/en/dev/internals/deprecation/#deprecation-removed-in-3-1
-
-.. automodule:: django_codemod.commands.django_31
-    :members:
-
-Django 3.2
-----------
-
-These codemodders should fix things `removed in Django 3.2`_.
-
-.. _removed in Django 3.2: https://docs.djangoproject.com/en/dev/internals/deprecation/#deprecation-removed-in-3-2
-
-.. automodule:: django_codemod.commands.django_32
+.. autoclass:: django_codemod.commands.Django30Command
     :members:
 
 Django 4.0
 ----------
 
-These codemodders should fix things `removed in Django 4.0`_.
+These command should fix things `removed in Django 4.0`_.
 
 .. _removed in Django 4.0: https://docs.djangoproject.com/en/dev/internals/deprecation/#deprecation-removed-in-4-0
 
-.. automodule:: django_codemod.commands.django_40
+.. autoclass:: django_codemod.commands.Django40Command
     :members:
